requirements.txt
scikit-learn==1.4.2 
imbalanced-learn==0.12.3 
xgboost==2.1.1
numpy
pandas
matplotlib
seaborn
fastapi
uvicorn
python-multipart
PyYAML
kagglehub
